simplify the ids
