+++ thirdeye-spi/src/main/java/ai/startree/thirdeye/spi/datalayer/dto/AbstractDTO.java
@@ -25,6 +25,7 @@ import org.apache.commons.lang3.builder.ToStringStyle;
 public abstract class AbstractDTO implements Serializable {
 
   private Long id;
+  private String namespace;
   private int version;
   private Timestamp createTime;
   private String createdBy;
@@ -37,6 +38,15 @@ public abstract class AbstractDTO implements Serializable {
 
   public AbstractDTO setId(final Long id) {
     this.id = id;
+    return this;
+  }
+
+  public String getNamespace() {
+    return namespace;
+  }
+
+  public AbstractDTO setNamespace(final String namespace) {
+    this.namespace = namespace;
     return this;
   }
